Implementation with Bulk Operations API
The latest Cosmos DB SDK introduced a new Bulk Operations API to do
up to 100 operations in a single request to the backend.

This commit adds this new API as the third import mechanism in addition
to the already existing two mechanisms.

This commit also updates dependencies to latest and refactors the
file and folder structure for additional clarity.